Move displayBirdSounds fn to DynamicSoundDisplay class & file
