+++ src/xenoCantoDynamicOutput.js
@@ -1,6 +1,16 @@
 import $ from 'jquery';
+import DataParsing from './data-parsing.js';
 
 export default class DynamicSoundDisplay {
+  static displayBirdSounds(response, commonName) {
+    $('#display-sounds').show();
+    $("#common-name").text(commonName);
+    const songsToOutput = DataParsing.filterForSongs(response);
+    const callsToOutput = DataParsing.filterForCalls(response);
+    $("#song").attr("src", songsToOutput[0].file);
+    $("#call").attr("src", callsToOutput[0].file);
+  }
+  
   static SongOutput() {
     let outputDiv = $("#songs-output");
     let htmlForSongOutput = `
